Rearrange sidebar folder structure
