+++ src/components/Sidebar/SidebarCards/MissionCard.vue
@@ -1,84 +1,55 @@
 <script setup lang="ts">
 import { Card, CardContent, CardTitle, CardFooter } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { computed } from "vue";
+import { computed, defineEmits } from "vue";
 import { Trash2 } from "lucide-vue-next";
 import { missionStore } from "@/lib/MissionStore";
-import { event } from "@tauri-apps/api";
 
-const props = defineProps<{
-  missionId: number;
+defineProps<{
+  missionNumber: number;
+  status: "Active" | "Inactive" | "Failed" | "Complete";
 }>();
 
 // Status Styles
-const statusStyles = {
+const statusStyles = computed(() => ({
   statusColor: {
     Inactive: "text-muted-foreground font-semibold",
     Failed: "text-destructive font-semibold",
     Active: "text-chart-4 font-semibold",
     Complete: "text-chart-2 font-semibold"
   }
-};
-
-const mission = missionStore.getMissionData(props.missionId);
-
-const handleZoneButtonClick = () => {
-  missionStore.setCurrentView("zone");
-  missionStore.setCurrentMissionID(props.missionId);
-};
-
-// Triggered on blur (unfocused) or enter key
-const handleMissionNameChange = (event: Event) => {
-  const newName = (event.target as HTMLInputElement).value;
-  missionStore.renameMission(props.missionId, newName);
-};
+}));
 </script>
 
 <template>
-  <Card v-if="mission" class="relative m-2 bg-sidebar-foreground p-2 text-foreground">
-    <!-- Mission Title -->
-    <CardTitle class="flex items-center gap-2">
-      <Input
-        @blur="handleMissionNameChange"
-        @keyup.enter="handleMissionNameChange"
-        v-if="mission.mission_status === 'Inactive'"
-        v-model="mission.mission_name"
-        class="flex-1"
+  <Card class="relative m-2 bg-sidebar-foreground p-2 text-foreground">
+    <!-- Trash Icon -->
+    <div class="absolute right-2 top-2 cursor-pointer">
+      <Trash2
+        @click="missionStore.view.deleteMission(missionNumber)"
+        class="h-5 w-5 text-secondary hover:text-destructive"
       />
-      <span v-else class="flex-1">
-        {{ mission.mission_name }}
-      </span>
+    </div>
 
-      <!-- Trash Icon -->
-      <div v-if="mission.mission_status == 'Inactive'" class="cursor-pointer">
-        <Trash2
-          @click.stop="missionStore.deleteMission(mission.mission_id)"
-          class="h-5 w-5 text-foreground hover:text-destructive"
-        />
-      </div>
-    </CardTitle>
+    <!-- Mission Title -->
+    <CardTitle class="text-x2 font-bold">Mission {{ missionNumber }}</CardTitle>
 
     <!-- Status Section -->
     <CardContent class="mt-2 flex flex-col items-start">
       <span class="font-semibold">
-        Status:
-        <span :class="statusStyles.statusColor[mission.mission_status]">{{
-          mission.mission_status
-        }}</span>
+        Status: <span :class="statusStyles.statusColor[status]">{{ status }}</span>
       </span>
     </CardContent>
 
     <!-- Submit Button -->
     <CardFooter class="mt-4 justify-start">
-      <Button 
-        class="mr-2" 
-        :disabled="mission.mission_status != 'Inactive'"
-        @click.stop="missionStore.startMission(mission.mission_id)"
+      <Button
+        class="mr-2"
+        :disabled="status === 'Active' || status === 'Complete' || status === 'Failed'"
       >
-        Start
+        Submit
       </Button>
-      <Button @click.stop="handleZoneButtonClick"> Zones </Button>
+      <Button> Zones </Button>
     </CardFooter>
   </Card>
 </template>
