+++ src/components/Sidebar/SidebarCards/VehicleCard.vue
@@ -1,80 +1,65 @@
 <script setup lang="ts">
-import { ref, watch, computed } from "vue";
-import { Card, CardContent, CardTitle, CardFooter } from "@/components/ui/card";
-import { Button } from "@/components/ui/button";
-import { Toggle } from "@/components/ui/toggle";
-import { VehicleEnum } from "@/lib/bindings";
-import { missionStore } from "@/lib/MissionStore";
-import { set } from "@vueuse/core";
+import { ref, computed } from 'vue'
+import { Card, CardContent, CardTitle, CardFooter } from '@/components/ui/card'
+import { Button } from '@/components/ui/button'
 
 // Define Props
-const props = defineProps<{
-  vehicleName: VehicleEnum;
-}>();
+defineProps<{
+  vehicleName: 'ERU' | 'MEA' | 'MRA'
+  patientStatus: 'Secured' | 'Unsecured'
+}>()
 
-const missionId = missionStore.view.currentMissionId;
+// Temp stages for the vehicle
+const stages = ['Search', 'Rescue', 'Transport', 'Completed']
 
-const vehicle =
-  missionId !== null ? missionStore.getVehicleData(missionId, props.vehicleName) : null;
+// Stage index
+const currentStageIndex = ref(0)
 
-const currentVehicleStage =
-  vehicle === undefined || vehicle === null || missionId === null
-    ? null
-    : vehicle.stages.find((stage) => stage.stage_id === vehicle.current_stage);
+// Computed property for current stage name
+const stageName = computed(() => stages[currentStageIndex.value])
 
-const patientStatusStyles = {
+// Status Styles
+const patientStatusStyles = computed(() => ({
   statusColor: {
-    Secured: "text-chart-2 font-semibold",
-    Unsecured: "text-destructive font-semibold"
+    'Secured': 'text-chart-2 font-semibold',
+    'Unsecured': 'text-destructive font-semibold'
   }
-};
+}))
+
+// Move to the next stage
+function nextStage() {
+  if (currentStageIndex.value < stages.length - 1) {
+    currentStageIndex.value++
+  }
+}
 </script>
 
 <template>
-  <Card
-    v-if="vehicle && missionId !== null"
-    class="relative m-2 bg-sidebar-foreground p-2 text-foreground"
-  >
+  <Card class="p-2 m-2 relative">
     <!-- Vehicle Name -->
     <CardTitle class="text-xl font-bold">{{ vehicleName }}</CardTitle>
 
     <!-- Vehicle Stage & Patient Status -->
     <CardContent class="mt-2 flex flex-col items-start">
-      <span
-        v-if="
-          vehicle.stages.length > 0 &&
-          (currentVehicleStage !== null || currentVehicleStage !== undefined)
-        "
-        class="font-semibold"
-      >
-        Stage: {{ currentVehicleStage?.stage_name }}
+      <span class="font-semibold">
+        Stage: {{ stageName }}
       </span>
-      <span v-else class="font-semibold">No Stages Available</span>
-
-      <!-- MEA & ERU are the only vehicles with patient status -->
-      <span v-if="vehicle.patient_status !== null" class="font-semibold">
-        Patient Status:
-        <span :class="patientStatusStyles.statusColor[vehicle.patient_status]">
-          {{ vehicle.patient_status }}
+      <span class="font-semibold">
+        Patient Status: 
+        <span :class="patientStatusStyles.statusColor[patientStatus]">
+          {{ patientStatus }}
         </span>
       </span>
     </CardContent>
 
     <!-- Next Stage Button -->
     <CardFooter class="mt-4 justify-start">
-      <!-- TODO: Add logic to disable if vehicle is at last stage or if mission is not submitted or not started-->
-      <Button
-        :disabled="vehicle.stages.length < 1 || vehicle.current_stage === vehicle.stages[vehicle.stages.length - 1].stage_id"
-        @click.stop="missionStore.transitionStage(missionId, props.vehicleName)">
+      <Button 
+        @click="nextStage"
+        :disabled="currentStageIndex === stages.length - 1"
+      >
         Next Stage
       </Button>
-      <span v-if="vehicle.is_auto !== null " class="font-semibold">
-        <Toggle 
-        class="bg-primary text-primary-foreground ml-2 hover:bg-primary/90 hover:text-primary-foreground"
-        @click.stop="missionStore.setAutoMode(missionId, props.vehicleName, !vehicle.is_auto)">
-        Auto 
-        </Toggle>
-      </span>
     </CardFooter>
   </Card>
 </template>
