+++ src/components/Sidebar/Tabs/StageView.vue
@@ -1,37 +1,62 @@
 <script setup lang="ts">
-import { computed, watch, reactive } from "vue";
-import { Button } from "@/components/ui/button";
+import { ref } from "vue";
 import StageCard from "@/components/Sidebar/SidebarCards/StageCard.vue";
-import { SidebarContent, SidebarFooter } from "@/components/ui/sidebar";
-import { Plus } from "lucide-vue-next";
-import { missionStore } from "@/lib/MissionStore";
 
-const currentMissionId = missionStore.view.currentMissionId;
-const currentVehicleName = missionStore.view.currentVehicleName;
-const stages = currentMissionId !== null && currentVehicleName !== null ? missionStore.getVehicleData(currentMissionId, currentVehicleName)?.stages : [];
+// Temporary stage data
+const stages = ref<
+  {
+    StageName: string;
+    StageStatus: "Incomplete" | "In Progress" | "Complete";
+    SearchArea: string[];
+  }[]
+>([
+  {
+    StageName: "Takeoff",
+    StageStatus: "Incomplete",
+    SearchArea: ["0 0", "0 1", "1 1", "1 0"]
+  },
+  {
+    StageName: "Search",
+    StageStatus: "Incomplete",
+    SearchArea: ["0 0", "0 1", "1 1", "1 0"]
+  },
+  {
+    StageName: "Landing",
+    StageStatus: "Incomplete",
+    SearchArea: ["0 0", "0 1", "1 1", "1 0"]
+  }
+]);
 
+// Function to add a new stage
+const addStage = () => {
+  const newStage = {
+    StageName: `Stage ${stages.value.length + 1}`,
+    StageStatus: "Incomplete" as const,
+    SearchArea: []
+  };
+  stages.value.push(newStage);
+};
+
+defineExpose({ addStage }); // Expose addStage to side bar
+
+// Function to delete a stage
+const deleteStage = (stageIndex: number) => {
+  stages.value.splice(stageIndex, 1);
+};
 </script>
 
 <template>
-  <SidebarContent class="bg-sidebar-background">
-    <div class="flex w-full flex-col items-center">
-      <div v-if="stages && stages.length > 0" class="w-full space-y-4">
-        <StageCard v-for="(stage, index) in stages" :key="index" :stageID="stage.stage_id" :stageIndex="index" />
-      </div>
-      <div v-else class="w-full text-center">
-        <span>No Stages Created</span>
-      </div>
-
+  <div class="flex w-full flex-col items-center">
+    <div v-if="stages.length > 0" class="w-full space-y-4">
+      <StageCard
+        v-for="(stage, index) in stages"
+        :key="index"
+        :missionNumber="index"
+        :stageName="stage.StageName"
+        :status="stage.StageStatus"
+        :searchArea="stage.SearchArea"
+        @deleteStage="deleteStage(index)"
+      />
     </div>
-  </SidebarContent>
-  <SidebarFooter class="bg-sidebar-background">
-    <Button class="flex flex-col items-center bg-transparent text-background shadow-none hover:bg-transparent" @click="
-      currentMissionId !== null &&
-      currentVehicleName !== null &&
-      missionStore.addStage(currentMissionId, currentVehicleName)
-      ">
-      <Plus class="h-5 w-5" />
-      Add Stage
-    </Button>
-  </SidebarFooter>
+  </div>
 </template>
