+++ src/components/Sidebar/Tabs/VehicleView.vue
@@ -1,21 +1,13 @@
 <script setup lang="ts">
 import VehicleCard from "@/components/Sidebar/SidebarCards/VehicleCard.vue";
-import { SidebarContent, SidebarGroup } from "@/components/ui/sidebar";
-import { VehicleEnum } from "@/lib/bindings";
-import { missionStore } from "@/lib/MissionStore";
-
-const handleClick = (vehicleName: VehicleEnum) => {
-  missionStore.setCurrentView("stage");
-  missionStore.setCurrentVehicleName(vehicleName);
-};
 </script>
 
 <template>
-  <SidebarContent class="bg-sidebar-background">
+  <div>
     <SidebarGroup>
-      <VehicleCard vehicleName="ERU" @click="handleClick('ERU')" />
-      <VehicleCard vehicleName="MEA" @click="handleClick('MEA')" />
-      <VehicleCard vehicleName="MRA" @click="handleClick('MRA')" />
+      <VehicleCard vehicleName="ERU" patientStatus="Secured" />
+      <VehicleCard vehicleName="MEA" patientStatus="Unsecured" />
+      <VehicleCard vehicleName="MRA" patientStatus="Unsecured" />
     </SidebarGroup>
-  </SidebarContent>
+  </div>
 </template>
